Refactoring
Renaming the package structure, where jdp stands for java design
patterns
